refactor: 총수익, 수익률 구하는 로직 Money 클래스에서 처리
