+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -30,24 +30,6 @@ class LottoResultTest {
                 .contains(Ranking.FOURTH);
     }
 
-    @Test
-    public void 로또_총수익_테스트1() {
-        PurchasedLotto purchasedLotto = createSamplePurchasedLotto();
-        Lotto winningLotto = new Lotto("1, 2, 3, 4, 5, 6");
-        LottoNo bonusNumber = new LottoNo(7);
-
-        LottoResult result = new LottoResult(purchasedLotto.compareLottos(winningLotto, bonusNumber));
-        BigDecimal actual = result.calculateWinningMoney();
-
-        List<Ranking> rankings = Arrays.asList(Ranking.FIRST, Ranking.SECOND, Ranking.THIRD, Ranking.FOURTH);
-        BigDecimal expected = rankings.stream()
-                .mapToInt(Ranking::getReward)
-                .mapToObj(BigDecimal::new)
-                .reduce(BigDecimal.ZERO, BigDecimal::add);
-
-        assertThat(actual).isEqualTo(expected);
-    }
-
     private PurchasedLotto createSamplePurchasedLotto() {
         List<Lotto> lottoList = Arrays.asList(
                 new Lotto(1, 2, 3, 4, 5, 6),
@@ -55,19 +37,5 @@ class LottoResultTest {
                 new Lotto(1, 2, 3, 4, 5, 8),
                 new Lotto(1, 2, 3, 4, 7, 8));
         return new PurchasedLotto(lottoList);
-    }
-
-    @Test
-    public void 로또_수익률_테스트() {
-        List<Ranking> rankings = Arrays.asList(Ranking.FIRST, Ranking.SECOND, Ranking.THIRD, Ranking.FOURTH);
-        LottoResult result = new LottoResult(rankings);
-        Money money = new Money(5000);
-
-        BigDecimal expectedWinningMoney = result.calculateWinningMoney();
-        BigDecimal divisor = new BigDecimal(money.getMoney());
-        BigDecimal expectedProfit = expectedWinningMoney.divide(divisor).setScale(2, RoundingMode.HALF_UP);
-        BigDecimal actualProfit = result.calculateWinningProfit(money);
-
-        assertThat(actualProfit).isEqualTo(expectedProfit);
     }
 }
